+++ tests/server/rpc/dslv1/test_lexer.py
